revert config upgrade
